update readme, changelog, contributing.md, add test comments, break 3.8
